requirement.txt
tornado
jinja2
MySQL-python
python-memcached
peewee==2.1.0
simplejson
PIL
